Refactor common utility module from core
Pulls out any private functions used as utilities in core into a
separate module.
